# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sidequestd API",
    "description": "Shared API for the Sidequestd web and mobile clients.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "The signed-in user's own record",
        "operationId": "read_current_user_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMe"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Edit your own profile",
        "description": "SPEC \u00a76.2. Omitted fields are left alone; an explicit null clears one.\n\nThe handle and email are not editable here \u2014 both are identity, and changing\nthem has consequences (dead profile links, a re-verification flow) that\nbelong in settings rather than in a profile edit form.",
        "operationId": "update_current_user_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMe"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/avatar": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Upload a profile picture",
        "description": "Store the image and point the profile at it (SPEC \u00a76.2).\n\nThe declared content type and filename are both ignored \u2014 the stored type is\nsniffed from the bytes and the key is server-generated. See\n`app.services.storage`.",
        "operationId": "upload_avatar_api_v1_users_me_avatar_put",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_avatar_api_v1_users_me_avatar_put"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMe"
                }
              }
            }
          },
          "413": {
            "description": "File exceeds the size limit"
          },
          "415": {
            "description": "Not an image we accept"
          },
          "503": {
            "description": "Object storage is unreachable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Remove your profile picture",
        "operationId": "delete_avatar_api_v1_users_me_avatar_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMe"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/favorites": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Your curated games, in pinned order",
        "operationId": "read_favorites_api_v1_users_me_favorites_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FavoriteGameEntry"
                  },
                  "type": "array",
                  "title": "Response Read Favorites Api V1 Users Me Favorites Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Reorder your pinned games",
        "operationId": "reorder_favorites_api_v1_users_me_favorites_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteReorder"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FavoriteGameEntry"
                  },
                  "type": "array",
                  "title": "Response Reorder Favorites Api V1 Users Me Favorites Put"
                }
              }
            }
          },
          "400": {
            "description": "Not a permutation of the current list"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Pin a game to your profile",
        "description": "Appends to the end. Returns the whole list so a client never has to guess\nthe resulting order.",
        "operationId": "add_favorite_api_v1_users_me_favorites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FavoriteGameEntry"
                  },
                  "type": "array",
                  "title": "Response Add Favorite Api V1 Users Me Favorites Post"
                }
              }
            }
          },
          "404": {
            "description": "No such game in the catalog"
          },
          "409": {
            "description": "Already pinned, or the list is full"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/favorites/{game_id}": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Unpin a game",
        "operationId": "remove_favorite_api_v1_users_me_favorites__game_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Catalog id of the pinned game",
              "title": "Game Id"
            },
            "description": "Catalog id of the pinned game"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FavoriteGameEntry"
                  },
                  "title": "Response Remove Favorite Api V1 Users Me Favorites  Game Id  Delete"
                }
              }
            }
          },
          "404": {
            "description": "Not one of your favorites"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/lists": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Your four backlog lists, each in your own order",
        "description": "SPEC \u00a76.9. All four lists unless `?status=` names one.\n\nUngated, unlike the endpoint below: this is your own backlog, and going\nprivate does not hide it from you.",
        "operationId": "read_own_lists_api_v1_users_me_lists_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BacklogStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return only this list instead of all four",
              "title": "Status"
            },
            "description": "Return only this list instead of all four"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacklogLists"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/username-available": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Check a handle before submitting the signup form",
        "description": "Advisory only \u2014 registration re-checks under the unique constraint.",
        "operationId": "check_username_api_v1_users_username_available_get",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Handle to check",
              "title": "Username"
            },
            "description": "Handle to check"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsernameAvailability"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{username}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "A public profile",
        "description": "SPEC \u00a76.2, readable signed-out.\n\nA private account still returns 200 with its shell and follower counts \u2014 that\nis what a stranger needs in order to send a follow request (SPEC \u00a76.7). What\nchanges is `can_view_content`: when it is false, `stats` is null and\n`favorite_games` is empty rather than the endpoint refusing.",
        "operationId": "read_profile_api_v1_users__username__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Public handle, case-insensitive",
              "title": "Username"
            },
            "description": "Public handle, case-insensitive"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "404": {
            "description": "No such user"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}/stats": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Profile stats",
        "description": "SPEC \u00a76.8. Every figure is zero (and the average null) until the reviews\nand backlog slices give it something to count.",
        "operationId": "read_stats_api_v1_users__user_id__stats_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "User id",
              "title": "User Id"
            },
            "description": "User id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileStats"
                }
              }
            }
          },
          "403": {
            "description": "Private account, viewer not approved"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{user_id}/backlog": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "This user's backlog lists",
        "description": "SPEC \u00a76.9: lists inherit account privacy, with no per-list toggle at MVP.\n\nSo this needs no rule of its own \u2014 `require_content_access` is the same gate\nthe reviews grid and the stats go through, which is what stops a backlog from\nbeing the one surface where a private account leaks.",
        "operationId": "read_user_lists_api_v1_users__user_id__backlog_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "User id",
              "title": "User Id"
            },
            "description": "User id"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BacklogStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return only this list instead of all four",
              "title": "Status"
            },
            "description": "Return only this list instead of all four"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacklogLists"
                }
              }
            }
          },
          "403": {
            "description": "Private account, viewer not approved"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserMe": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "email_verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Verified At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "display_name",
          "bio",
          "avatar_url",
          "is_private",
          "created_at",
          "email",
          "email_verified_at"
        ],
        "title": "UserMe",
        "description": "The authenticated user's own record \u2014 adds fields that are private to them."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /auth/login",
        "scheme": "bearer"
      }
    }
  }
}